test: fixture 분리.
