+++ batch/src/test/java/kr/co/techcourse/prolog/batch/configuration/schedule/ScheduleBatchFinderTest.java
@@ -3,16 +3,11 @@ package kr.co.techcourse.prolog.batch.configuration.schedule;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kr.co.techcourse.prolog.batch.job.common.EnableBatchJob;
+import kr.co.techcourse.prolog.batch.configuration.schedule.fixture.TestConfiguration_1;
+import kr.co.techcourse.prolog.batch.configuration.schedule.fixture.TestConfiguration_2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobParametersIncrementer;
-import org.springframework.batch.core.JobParametersValidator;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
 class ScheduleBatchFinderTest {
 
@@ -34,81 +29,17 @@ class ScheduleBatchFinderTest {
             .ignoringFields("jobBean")
             .isEqualTo(List.of(
                 new BatchJob(
-                    "testcron",
+                    "* * * * * *",
                     new String[]{"a:a", "b:b"},
                     ':',
                     null
                 ),
                 new BatchJob(
-                    "testcron",
+                    "* * * * * *",
                     new String[]{},
                     '=',
                     null
                 )
             ));
-    }
-
-    @Configuration
-    public static class TestConfiguration_1 {
-
-        @Bean
-        public Job test1() {
-            return emptyJob();
-        }
-
-        @EnableBatchJob(cron = "testcron", jobParameters = {"a:a", "b:b"}, delimiter = ':')
-        @Bean("testJob2")
-        public Job test2() {
-            return emptyJob();
-        }
-
-        @Bean("testJob3")
-        public Job test3() {
-            return emptyJob();
-        }
-    }
-
-    @Configuration
-    public static class TestConfiguration_2 {
-
-        @EnableBatchJob(cron = "testcron")
-        @Bean
-        public Job test4() {
-            return emptyJob();
-        }
-
-        @Bean("testJob5")
-        public Job test5() {
-            return emptyJob();
-        }
-    }
-
-    public static Job emptyJob() {
-        return new Job() {
-            @Override
-            public String getName() {
-                return null;
-            }
-
-            @Override
-            public boolean isRestartable() {
-                return false;
-            }
-
-            @Override
-            public void execute(JobExecution execution) {
-
-            }
-
-            @Override
-            public JobParametersIncrementer getJobParametersIncrementer() {
-                return null;
-            }
-
-            @Override
-            public JobParametersValidator getJobParametersValidator() {
-                return null;
-            }
-        };
     }
 }
